+++ plotly/api/v2/__init__.py
@@ -1,0 +1,4 @@
+from __future__ import absolute_import
+
+from plotly.api.v2 import (files, folders, grids, images, plot_schema, plots,
+                           users)