+++ plotly/offline/offline.py
@@ -5,7 +5,6 @@
 """
 from __future__ import absolute_import
 
-import json
 import os
 import uuid
 import warnings
@@ -13,22 +12,15 @@ from pkg_resources import resource_string
 import time
 import webbrowser
 
+from requests.compat import json as _json
+
 import plotly
-from plotly import tools, utils
+from plotly import optional_imports, tools, utils
 from plotly.exceptions import PlotlyError
 
-try:
-    import IPython
-    from IPython.display import HTML, display
-    _ipython_imported = True
-except ImportError:
-    _ipython_imported = False
-
-try:
-    import matplotlib
-    _matplotlib_imported = True
-except ImportError:
-    _matplotlib_imported = False
+ipython = optional_imports.get_module('IPython')
+ipython_display = optional_imports.get_module('IPython.display')
+matplotlib = optional_imports.get_module('matplotlib')
 
 __PLOTLY_OFFLINE_INITIALIZED = False
 
@@ -111,7 +103,7 @@ def init_notebook_mode(connected=False):
     your notebook, resulting in much larger notebook sizes compared to the case
     where `connected=True`.
     """
-    if not _ipython_imported:
+    if not ipython:
         raise ImportError('`iplot` can only run inside an IPython Notebook.')
 
     global __PLOTLY_OFFLINE_INITIALIZED
@@ -148,7 +140,7 @@ def init_notebook_mode(connected=False):
             '</script>'
             '').format(script=get_plotlyjs())
 
-    display(HTML(script_inject))
+    ipython_display.display(ipython_display.HTML(script_inject))
     __PLOTLY_OFFLINE_INITIALIZED = True
 
 
@@ -183,10 +175,12 @@ def _plot_html(figure_or_data, config, validate, default_width,
         height = str(height) + 'px'
 
     plotdivid = uuid.uuid4()
-    jdata = json.dumps(figure.get('data', []), cls=utils.PlotlyJSONEncoder)
-    jlayout = json.dumps(figure.get('layout', {}), cls=utils.PlotlyJSONEncoder)
+    jdata = _json.dumps(figure.get('data', []), cls=utils.PlotlyJSONEncoder)
+    jlayout = _json.dumps(figure.get('layout', {}),
+                          cls=utils.PlotlyJSONEncoder)
     if 'frames' in figure_or_data:
-        jframes = json.dumps(figure.get('frames', {}), cls=utils.PlotlyJSONEncoder)
+        jframes = _json.dumps(figure.get('frames', {}),
+                              cls=utils.PlotlyJSONEncoder)
 
     configkeys = (
         'editable',
@@ -211,7 +205,7 @@ def _plot_html(figure_or_data, config, validate, default_width,
     )
 
     config_clean = dict((k, config[k]) for k in configkeys if k in config)
-    jconfig = json.dumps(config_clean)
+    jconfig = _json.dumps(config_clean)
 
     # TODO: The get_config 'source of truth' should
     # really be somewhere other than plotly.plotly
@@ -330,7 +324,7 @@ def iplot(figure_or_data, show_link=True, link_text='Export to plot.ly',
             'plotly.offline.init_notebook_mode() '
             '# run at the start of every ipython notebook',
         ]))
-    if not tools._ipython_imported:
+    if not ipython:
         raise ImportError('`iplot` can only run inside an IPython Notebook.')
 
     config = {}
@@ -341,7 +335,7 @@ def iplot(figure_or_data, show_link=True, link_text='Export to plot.ly',
         figure_or_data, config, validate, '100%', 525, True
     )
 
-    display(HTML(plot_html))
+    ipython_display.display(ipython_display.HTML(plot_html))
 
     if image:
         if image not in __IMAGE_FORMATS:
@@ -357,7 +351,7 @@ def iplot(figure_or_data, show_link=True, link_text='Export to plot.ly',
         # allow time for the plot to draw
         time.sleep(1)
         # inject code to download an image of the plot
-        display(HTML(script))
+        ipython_display.display(ipython_display.HTML(script))
 
 
 def plot(figure_or_data, show_link=True, link_text='Export to plot.ly',
@@ -687,7 +681,7 @@ def enable_mpl_offline(resize=False, strip_style=False,
     """
     init_notebook_mode()
 
-    ip = IPython.core.getipython.get_ipython()
+    ip = ipython.core.getipython.get_ipython()
     formatter = ip.display_formatter.formatters['text/html']
     formatter.for_type(matplotlib.figure.Figure,
                        lambda fig: iplot_mpl(fig, resize, strip_style, verbose,