+++ plotly/tests/test_optional/test_offline/test_offline.py
@@ -6,16 +6,16 @@ from __future__ import absolute_import
 
 from nose.tools import raises
 from nose.plugins.attrib import attr
+from requests.compat import json as _json
 
 from unittest import TestCase
-import json
 
 import plotly
+from plotly import optional_imports
 
-# TODO: matplotlib-build-wip
-from plotly.tools import _matplotlylib_imported
+matplotlylib = optional_imports.get_module('plotly.matplotlylib')
 
-if _matplotlylib_imported:
+if matplotlylib:
     import matplotlib
     # Force matplotlib to not use any Xwindows backend.
     matplotlib.use('Agg')
@@ -34,7 +34,6 @@ class PlotlyOfflineTestCase(TestCase):
         plotly.offline.iplot([{}])
 
     def test_iplot_works_after_you_call_init_notebook_mode(self):
-        plotly.tools._ipython_imported = True
         plotly.offline.init_notebook_mode()
         plotly.offline.iplot([{}])
 
@@ -47,7 +46,6 @@ class PlotlyOfflineTestCase(TestCase):
         y = [100, 200, 300]
         plt.plot(x, y, "o")
 
-        plotly.tools._ipython_imported = True
         plotly.offline.init_notebook_mode()
         plotly.offline.iplot_mpl(fig)
 
@@ -75,8 +73,8 @@ class PlotlyOfflineMPLTestCase(TestCase):
         figure = plotly.tools.mpl_to_plotly(fig)
         data = figure['data']
         layout = figure['layout']
-        data_json = json.dumps(data, cls=plotly.utils.PlotlyJSONEncoder)
-        layout_json = json.dumps(layout, cls=plotly.utils.PlotlyJSONEncoder)
+        data_json = _json.dumps(data, cls=plotly.utils.PlotlyJSONEncoder)
+        layout_json = _json.dumps(layout, cls=plotly.utils.PlotlyJSONEncoder)
         html = self._read_html(plotly.offline.plot_mpl(fig))
 
         # just make sure a few of the parts are in here