+++ plotly/utils.py
@@ -7,7 +7,6 @@ Low-level functionality NOT intended for users to EVER use.
 """
 from __future__ import absolute_import
 
-import json
 import os.path
 import re
 import sys
@@ -15,27 +14,16 @@ import threading
 import decimal
 
 import pytz
+from requests.compat import json as _json
 
+from plotly.optional_imports import get_module
 
 from . exceptions import PlotlyError
 
-try:
-    import numpy
-    _numpy_imported = True
-except ImportError:
-    _numpy_imported = False
-
-try:
-    import pandas
-    _pandas_imported = True
-except ImportError:
-    _pandas_imported = False
-
-try:
-    import sage.all
-    _sage_imported = True
-except ImportError:
-    _sage_imported = False
+# Optional imports, may be None for users that only use our core functionality.
+numpy = get_module('numpy')
+pandas = get_module('pandas')
+sage_all = get_module('sage.all')
 
 
 ### incase people are using threading, we lock file reads
@@ -51,7 +39,7 @@ def load_json_dict(filename, *args):
         lock.acquire()
         with open(filename, "r") as f:
             try:
-                data = json.load(f)
+                data = _json.load(f)
                 if not isinstance(data, dict):
                     data = {}
             except:
@@ -66,7 +54,7 @@ def save_json_dict(filename, json_dict):
     """Save json to file. Error if path DNE, not a dict, or invalid json."""
     if isinstance(json_dict, dict):
         # this will raise a TypeError if something goes wrong
-        json_string = json.dumps(json_dict, indent=4)
+        json_string = _json.dumps(json_dict, indent=4)
         lock.acquire()
         with open(filename, "w") as f:
             f.write(json_string)
@@ -112,7 +100,7 @@ class NotEncodable(Exception):
     pass
 
 
-class PlotlyJSONEncoder(json.JSONEncoder):
+class PlotlyJSONEncoder(_json.JSONEncoder):
     """
     Meant to be passed as the `cls` kwarg to json.dumps(obj, cls=..)
 
@@ -149,7 +137,8 @@ class PlotlyJSONEncoder(json.JSONEncoder):
         #    1. `loads` to switch Infinity, -Infinity, NaN to None
         #    2. `dumps` again so you get 'null' instead of extended JSON
         try:
-            new_o = json.loads(encoded_o, parse_constant=self.coerce_to_strict)
+            new_o = _json.loads(encoded_o,
+                                parse_constant=self.coerce_to_strict)
         except ValueError:
 
             # invalid separators will fail here. raise a helpful exception
@@ -158,10 +147,10 @@ class PlotlyJSONEncoder(json.JSONEncoder):
                 "valid JSON separators?"
             )
         else:
-            return json.dumps(new_o, sort_keys=self.sort_keys,
-                              indent=self.indent,
-                              separators=(self.item_separator,
-                                          self.key_separator))
+            return _json.dumps(new_o, sort_keys=self.sort_keys,
+                               indent=self.indent,
+                               separators=(self.item_separator,
+                                           self.key_separator))
 
     def default(self, obj):
         """
@@ -210,7 +199,7 @@ class PlotlyJSONEncoder(json.JSONEncoder):
                 return encoding_method(obj)
             except NotEncodable:
                 pass
-        return json.JSONEncoder.default(self, obj)
+        return _json.JSONEncoder.default(self, obj)
 
     @staticmethod
     def encode_as_plotly(obj):
@@ -231,12 +220,12 @@ class PlotlyJSONEncoder(json.JSONEncoder):
     @staticmethod
     def encode_as_sage(obj):
         """Attempt to convert sage.all.RR to floats and sage.all.ZZ to ints"""
-        if not _sage_imported:
+        if not sage_all:
             raise NotEncodable
 
-        if obj in sage.all.RR:
+        if obj in sage_all.RR:
             return float(obj)
-        elif obj in sage.all.ZZ:
+        elif obj in sage_all.ZZ:
             return int(obj)
         else:
             raise NotEncodable
@@ -244,7 +233,7 @@ class PlotlyJSONEncoder(json.JSONEncoder):
     @staticmethod
     def encode_as_pandas(obj):
         """Attempt to convert pandas.NaT"""
-        if not _pandas_imported:
+        if not pandas:
             raise NotEncodable
 
         if obj is pandas.NaT:
@@ -255,7 +244,7 @@ class PlotlyJSONEncoder(json.JSONEncoder):
     @staticmethod
     def encode_as_numpy(obj):
         """Attempt to convert numpy.ma.core.masked"""
-        if not _numpy_imported:
+        if not numpy:
             raise NotEncodable
 
         if obj is numpy.ma.core.masked: